refactor: instructor integration
